Write integration test for destroy_survey_template

+++ app/helpers/survey_templates_helper.rb
@@ -7,7 +7,7 @@ module SurveyTemplatesHelper
   #
   def action_buttons(survey_template)
     edit_link = link_to 'Edit', edit_survey_template_path(survey_template), class: 'btn btn-primary'
-    delete_link = link_to 'Delete', survey_template_path(survey_template), class: 'btn btn-danger', method: :delete
+    delete_link = link_to 'Delete', survey_template_path(survey_template), class: 'btn btn-danger', method: :delete, data: { confirm: 'Are you sure?' }
 
     edit_link + delete_link
   end
